ocrbridge/requirements.txt
google-generativeai>=0.3.0
python-telegram-bot>=20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
streamlit>=1.28.0
requests>=2.31.0